backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
anthropic==0.28.0
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
youtube-transcript-api==0.6.2
pydantic==2.7.1
faster-whisper==1.1.1
yt-dlp==2025.3.31
twscrape==0.14.1
